move digest extraction from sc-consensus to sp-consensus
